Moving PushAction Receiver code to a common util function
Trampoline Activity now performs all the actions previously carried out by PushActionReceiver using a util function where all the methods are migrated.

Pending - PushActionReceiver to now handle only those conditions where the app knows if the Activity is not going to open. i.e CustomAction Handling with open App flag turned false
